PLM Report Automation using Selenium
I wrote these codes to help a co-worker to automate his repetitive work. He was downloading the same two reports from the PLM system every day. Since the size of the reports was so large, it tooks almost an hour to download and he couldn't do anything else on his computer. Now, thanks to these codes, the reports are ready before even he gets to work.